debug message

+++ serial.cpp
@@ -70,7 +70,7 @@ inline int calc_blocked_index(double x, double y, int nbins, int nblocks, int bi
 }
 
 // Initialize Simulation
-void init_simulation(particle_t* __restrict parts, int num_parts, double size, double bin_size_param, double block_size_param) {
+void init_simulation(particle_t* __restrict parts, int num_parts, double size, double bin_size_param, int block_size_param) {
     // Input validation
     if (size <= 0 || bin_size_param <= 0 || block_size_param <= 0 || num_parts <= 0 || !parts) {
         std::cerr << "Invalid parameters: size=" << size
